Refactoring FTs.
