+++ functional_tests/test_layout_and_styling.py
@@ -1,0 +1,25 @@
+from selenium.webdriver.common.keys import Keys
+from .base import FunctionalTest
+
+class LayoutAndStylingTest(FunctionalTest):
+
+	def test_layout_and_style(self):
+		self.browser.get(self.live_server_url)
+		self.browser.set_window_size(1024, 768)
+
+		inputbox = self.browser.find_element_by_id('id_new_item')
+		self.assertAlmostEqual(
+			inputbox.location['x'] + inputbox.size['width'] / 2,
+			512, delta=10
+			)
+
+		inputbox.send_keys('Testing')
+		inputbox.send_keys(Keys.ENTER)
+
+		self.wait_for_row_in_table('1: Testing')
+
+		inputbox = self.browser.find_element_by_id('id_new_item')
+		self.assertAlmostEqual(
+			inputbox.location['x'] + inputbox.size['width'] / 2,
+			512, delta=10
+			)
